Change overlays shown as user clicks on day
Moved maps js into a new file.
