+++ manyfews/package.json
@@ -20,6 +20,7 @@
     "webpack-cli": "^4.9.2"
   },
   "dependencies": {
-    "bootstrap": "^5.1.3"
+    "bootstrap": "^5.1.3",
+    "jquery": "^3.6.0"
   }
 }
